Added option to extract faces or just detect faces

+++ main.py
@@ -6,6 +6,16 @@ import os
 rootdir = sys.argv[1]
 fd = FaceDetector()
 
+extract = False
+
+# Check to see if a second argument
+# was entered, if so see if it was
+# the extract flag
+if len(sys.argv) > 2:
+    if sys.argv[2].lower() == "--extract":
+        extract = True
+
+
 # List of extensions so search for
 extensions = ["jpg", "png", "gif", "jpeg"]
 
@@ -17,5 +27,7 @@ for subdir, dirs, files in os.walk(rootdir):
         # Check if they are images
         for extension in extensions:
             if extension in f:
-                detected = fd.detect(os.path.join(subdir, f))
-                os.remove(os.path.join(subdir, f))
+                detected = fd.detect(os.path.join(subdir, f), extract)
+
+                if not detected or extract:
+                    os.remove(os.path.join(subdir, f))
